+++ R/plot.wfdb.R
@@ -1,4 +1,17 @@
-#' The function displays ECG signals in a layout corresponding to standard paper ECG printouts
+#' The function displays WFDB signals in a layout corresponding to standard paper ECG printouts
+#'
+#' ECG signals are read from files in WFDB format.
+#'
+#' WFDB (WaveForm DataBase) is a standard file format for storing, reading,
+#' and analyzing physiological time-series signals. It is widely used for
+#' signals such as ECG, EEG, blood pressure, respiration, and other biomedical
+#' waveforms. It is the file format used by the PhysioNet project and is commonly
+#' used in research datasets.
+#'
+#' A WFDB record typically consists of two main files:
+#' \code{.dat} - binary signal samples (waveform values), and \code{.hea} - a header
+#' file describing how to interpret the data. In some cases, additional annotation
+#' files such as \code{.atr} may be present, containing beat labels or rhythm annotations.
 #'
 #' A typical ECG paper layout was used, with a small grid of 0.04 s × 0.1 mV and a
 #' large grid of 0.20 s × 0.5 mV.
@@ -7,7 +20,7 @@
 #' @importFrom graphics lines segments
 #' @importFrom stats median
 #'
-#' @param x Object of class \code{ecg} (from \code{read_ecg_signals()}).
+#' @param x Object of class \code{wfdb} (from \code{read_wfdb_signals()}).
 #'
 #' @param begin Time point (in seconds) at which to start plotting.
 #'
@@ -33,7 +46,7 @@
 #' @examples
 #' # ECG data comes from https://physionet.org/content/ptb-xl/1.0.3/
 #' file <- system.file("extdata", "00001_lr.hea", package = "MatchingPursuit")
-#' out <- read_ecg_signals(file)
+#' out <- read_wfdb_signals(file)
 #'
 #' plot(
 #'   x = out,
@@ -44,7 +57,7 @@
 #'   small_squares = TRUE
 #' )
 #'
-plot.ecg <- function(
+plot.wfdb <- function(
     x,
     begin,
     end,
@@ -62,34 +75,34 @@ plot.ecg <- function(
   old.par <- par(no.readonly = TRUE)
   on.exit(par(old.par), add = FALSE)
 
-  if (!inherits(x, "ecg")) {
-    stop("'x' must be an object of class 'ecg'.")
+  if (!inherits(x, "wfdb")) {
+    stop("'x' must be an object of class 'wfdb'.")
   }
 
-  ecg <- as.matrix(x$signal)
+  wfdb <- as.matrix(x$signal)
   sf <- x$sampling_frequency
-  channels <- ncol(ecg)
+  channels <- ncol(wfdb)
 
   main <- paste("record name: ", x$record_name, sep = "")
 
   # Each column is centered around its median. In signals like ECG/EGM, this
   # helps remove the base-level offset (DC offset), making channels more comparable.
   # Following this line, each channel has a median of approximately zero.
-  md <- apply(ecg, 2, median)
-  ecg <- sweep(ecg, 2, md, "-")
+  md <- apply(wfdb, 2, median)
+  wfdb <- sweep(wfdb, 2, md, "-")
 
   from <- begin * sf
   to <- end * sf
 
-  ecg <- ecg[from:to, ]
+  wfdb <- wfdb[from:to, ]
 
-  lead.names <- colnames(ecg)
-  n <- nrow(ecg)
+  lead.names <- colnames(wfdb)
+  n <- nrow(wfdb)
 
   # time points
   t <- seq(begin, by = 1 / sf, length.out = n)
 
-  # duration of ECG signal (in sec.)
+  # duration of wfdb signal (in sec.)
   duration <- n / sf
 
   # panel_height - single strip height (mV)
@@ -142,7 +155,7 @@ plot.ecg <- function(
     if (zero_line) segments(begin, y0, end, y0, col = "blue", lwd = 0.5)
 
     ## signal
-    lines(t, ecg[, i] + y0, lwd = 1)
+    lines(t, wfdb[, i] + y0, lwd = 1)
 
     ## lead names
     shift <- (end - begin) * 0.02